pyproject.toml
[tool.poetry]
name = "fastapi-pubsub"
version = "0.1.0"
description = ""
authors = ["j-dsouza <james.dsouza@outlook.com>"]
readme = "README.md"
packages = [{include = "fastapi_pubsub"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
google-cloud-pubsub = "^2.15.2"
uvicorn = "^0.21.1"
typer = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
